cloud/lambdas/transcription-worker/requirements.txt
# REM Transcription Worker Lambda Dependencies

# AWS SDK
boto3>=1.34.0

# Whisper transcription (use openai-whisper instead of faster-whisper to avoid PyAV dependency)
openai-whisper>=20231117

# OpenAI for embeddings and summarization
openai>=1.0.0

# Speaker diarization
pyannote.audio>=3.1.0

# Audio processing
torch>=2.0.0
torchaudio>=2.0.0

# Local development
python-dotenv>=1.0.0
